Switch to 64-bit ISD-ASes (#1453)
Context: https://github.com/scionproto/scion/issues/1451

Also:
- Include raw scion packet in wireshark plugin, to allow for easier
  extraction.
- Fix some bugs in hpkt.WriteScnPkt:
  - L4 payload length wasn't set.
  - regBaseHeaderLen was incorrectly calculated, and only accidentally
    correct until now.
- Add ./tools/pktcap to save raw scion packets from the network to a
  file.
- Use real scion packets for go/border/rpkt and go/lib/hpkt unit tests,
  and document how to generate/capture them.
- Remove ancient code from libscion that dates from the pre-capnp era.
- Make hpkt's write test a lot less painful to maintain, by having it
  write and then parse the packet, rather than having to hand edit a
  binary representation.
- Use strconv.ParseUint instead of strconv.Atoi when parsing ISDs in Go,
  for greater control.
- Update various go data structures to use the new addr.ISD/adddr.AS
  types (e.g. TRCs and Certificates)
- Update libscion's extension.h and scion.lua for the new HBH extn
  numbers.
- Update SCMP RecordPath for 64-bit ISDASes
- Add some extra sanity checks for SCMP RecordPath.
- Minor cleanup of go/tools/scmp error reporting, so it's more obvious
  what part is the error.
- Include the types when the l4 or payload are not the expected type in
  scmp tool.

+++ proto/asm_exts.capnp
@@ -7,7 +7,7 @@ struct RoutingPolicyExt{
     set @0 :Bool;   # Is the extension present? Every extension must include this field.
     polType @1 :UInt8;  # The policy type
     ifID @2 :UInt64;
-    isdases @3 :List(UInt32);
+    isdases @3 :List(UInt64);
 }
 
 struct ISDAnnouncementExt{
